+++ src/test/OracleWrapper.sol
@@ -4,12 +4,12 @@ pragma solidity ^0.8.4;
 import {Oracle} from "../Oracle.sol";
 
 contract OracleWrapper is Oracle {
-    function writeWrapper(uint256 protoId, uint128 price) external {
+    function writeWrapper(uint protoId, uint128 price) external {
         Oracle.write(protoId, price);
     }
 
     function observeWrapper(
-        uint256 protoId,
+        uint protoId,
         uint128[] calldata secondsAgos,
         uint128 curWorth
     ) external view returns (uint128[] memory cumulativePrices) {
